Hyunsoo

+++ fullStack/src/main/java/capstone/fullstack/repository/local/commerce/SalesRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface SalesRepository extends JpaRepository<Sales, Long> {
     //기본 CRUD는 그냥 사용
 
-    List<Sales> findByCommercialCodeAndServiceName(Long commercialCode, String serviceName);
+    List<Sales> findByCommercialCodeAndServiceName(Integer commercialCode, String serviceName);
 
     List<Sales> findByCommercialCodeInAndServiceNameEquals(List<Integer> commercialCodeList, String serviceName);  //코드 하나라도 일치하면서 서비스 업종명 동일.
